requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
celery==5.3.4
redis==5.0.1
openai==1.3.0
python-dotenv==1.0.0
requests
langchain_openai
langgraph
pydantic_settings